👌 IMPROVE: Filter PRs by checking for commits in branch
If a branch is specified PRs are filtered by checking their PR base branch.
This fails to pick up all PRs when multiple PRs are merged into one branch, followed by that branch being merged into the mainline via a single PR.

If the `--since` argument is a git ref (_not_ a date) this is used to get a list of commits in the requested branch after `--since`. The list of PRs is then filtered by checking whether the PR commit appears in this list of branch commits. If the commit history for merged PRs/branches is maintained this should ensure PRs merged into one branch that are subsequently merged altogether into another branch are picked up.

Closes https://github.com/executablebooks/github-activity/issues/50